display more user friendly error
proejct_name is the only required field that could be flagged (delivery/technical contact are already accounted for) so this fix works with the current system

any other required fields would have to be accounted for. previously we have mapped this using a dict, but there is no need for it currently.