KEH-528 - Ability to select none (#14)
* button text unhidden

* add second button on /survey

* clean formScripts logic

* fallback when trying to retrieve data

* project.stage fix

* remove certain check logic

delivery manager or technical contact and project name are needed

* align icons in accordion

* minor changes to /survey logic

* view_project is dynamically rendered with JS

select none means missing data may occur and nunjucks doesnt have a good workaround for that

* rm print

* update validation check for languages

* display more user friendly error

proejct_name is the only required field that could be flagged (delivery/technical contact are already accounted for) so this fix works with the current system

any other required fields would have to be accounted for. previously we have mapped this using a dict, but there is no need for it currently.

* rename primary and secondary btns

* remove validation in summary pages

* remove repeated height styling code

* fix None error when loading edit page

* reformatting certain areas

* fix post py code

* fix formatting and loading errors

* rm rogue print

* fix render data and save data

* rm print

* render project description on validate details

* validation on render source control links

* fallback for doc link submission

* log errors

* add escapeHTML to project summary

* lint application.py

* rm unused test

* add linting for tests folder

* lint tests and update create project

* sanitise project_name

* fix developed logic

* change developed input ids

* major reformat

Take JS out of validate_details and put in separate JS, same for summary pages.

Fix my misunderstanding of how developed was stored.

It is stored as a list of strings ["", ""]

list[0] is In-house/Outsourced/Partnership then the list[1] is either empty or the other company

* change button ids

* escapeHTML content

* fix displaying HTML

* fix TL and DM display on project summary

* fallback when getting SecretString

* more detailed exception

* function to build p element

* view_project js to external file

* developed to external js file

* rm legacy format for developed obj

* Update test_project_creation.py

https://random-word-api.herokuapp.com/word returns:

java.lang.NoClassDefFoundError: Could not initialize class play.api.libs.json.Writes$

so I switched the random word API provider

* small cleanup

* testing choice GCP

* project_summary fix

* developed showing when empty